test: Small refactoring of serialization tests.
